Seprated alingment macros from geo file and started to make change to take geo parameters from the EPICS
