+++ startup/50-staff.py
@@ -1,0 +1,51 @@
+def phi_track(alpha_ini, alpha_stop, num_alpha):
+    # for eta in range(eta_start, eta_stop, nb_eta):
+    # eta
+
+    dif = np.zeros((2, num_alpha))
+    for i, alpha in enumerate(range(0, num_alpha, 1)):
+        alpha_re = alpha_ini + (i * (alpha_stop - alpha_ini) / num_alpha)
+        print(i, alpha_ini, alpha_re)
+        yield from bps.mv(geo, alpha_re)
+        yield from bp.rel_scan([quadem], geo.phi, -0.01, 0.01, 40)
+        print(peaks.cen["quadem_current2_mean_value"] - geo.ca(alpha_re).phi)
+        dif[0, i] = alpha_re
+
+        dif[1, i] = peaks.cen["quadem_current2_mean_value"] - geo.ca(alpha_re).phi
+
+    print(dif)
+    import matplotlib.pyplot as plt
+
+    plt.figure()
+    plt.plot(dif[0, :], dif[1, :])
+    plt.show()
+
+
+def ih_track(alpha_ini, alpha_stop, num_alpha):
+    # for ih in range(alpha_ini,alpha_stop, nb_alpha):
+
+    dif = np.zeros((3, num_alpha))
+    for i, alpha in enumerate(range(0, num_alpha, 1)):
+        alpha_re = alpha_ini + (i * (alpha_stop - alpha_ini) / num_alpha)
+        print(i, alpha_ini, alpha_re)
+        yield from bps.mv(geo, alpha_re)
+        yield from bp.rel_scan([quadem], geo.phi, -0.010, 0.010, 20)
+        yield from bps.mv(geo.phi, peaks.cen["quadem_current2_mean_value"])
+        dif[2, i] = peaks.cen["quadem_current2_mean_value"] - geo.ca(alpha_re).phi
+        yield from bp.rel_scan([quadem], geo.ih, -0.5, 0.5, 20)
+        # is the next line corrct, geo.ca.
+        print(peaks.cen["quadem_current3_mean_value"] - geo.ca(alpha_re).ih)
+        dif[0, i] = alpha_re
+        dif[1, i] = peaks.cen["quadem_current3_mean_value"] - geo.ca(alpha_re).ih
+
+    print(dif)
+    import matplotlib.pyplot as plt
+
+    plt.figure()
+    plt.subplot(211)
+    plt.title("IH_track")
+    plt.plot(dif[0, :], dif[1, :])
+    plt.subplot(212)
+    plt.title("Phi_track")
+    plt.plot(dif[0, :], dif[2, :])
+    plt.show()
